requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
diffusers==0.32.2
distro==1.9.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
openai==1.59.8
packaging==24.2
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
python-pptx==1.0.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
XlsxWriter==3.2.0
zipp==3.21.0
